% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScPoApps-package.R
\docType{package}
\name{ScPoApps}
\alias{ScPoApps}
\title{ScPoApps: Teaching Econometrics with R at SciencesPo}
\description{
The is the R package for the 2nd year UG
course in econometrics at SciencesPo.
}
\section{Tutorials}{

The package contains several tutorials which can be
launched via \code{\link[=runTutorial]{runTutorial()}}
}

\section{Apps}{

The package also contains several interactive apps
which you launch via \code{\link[=launchApp]{launchApp()}}
}

